work on "public" ingestion

+++ importer/src/main/scala/uk/org/nbn/nbnv/importer/ingestion/PublicIngester.scala
@@ -16,45 +16,46 @@ class PublicIngester @Inject()(log: Logger, em: EntityManager, repo: Repository,
 
     def update(p: TaxonObservationPublic) {
 
-      // set the fields which are always the same in the public record
+      // set the fields which are always the same as the main record
+      p.setObservationID(o.getObservationID)
       p.setSampleID(o.getSampleID)
       p.setObservationKey(o.getObservationKey)
 
       // set the site to that of the observation, or set it null
-      // val site = if (metadata.siteIsPublic) o.getSiteID else null
-      // p.setSiteID(site)
-      //
+      val site = if (metadata.siteIsPublic) o.getSiteID else null
+      p.setSiteID(site)
+
       // set the recorder & determiner to those of the observation, or set them null
-      // val recorder = if (metadata.recorderAndDeterminerArePublic) o.getRecorderID else null
-      // val determiner = if (metadata.recorderAndDeterminerArePublic) o.getDeterminerID else null
-      // p.setRecorderID(recorder)
-      // p.setDeterminerID(determiner)
-      //
+      val recorder = if (metadata.recorderAndDeterminerArePublic) o.getRecorderID else null
+      val determiner = if (metadata.recorderAndDeterminerArePublic) o.getDeterminerID else null
+      p.setRecorderID(recorder)
+      p.setDeterminerID(determiner)
+
       // set the feature to a potentially parent grid square feature
       // p.setFeatureID(feature)
     }
 
     repo.getTaxonObservationPublic(o.getObservationID) match {
       case Some(p) => {
-        // delete the public record if it's now sensitive
-        // if (metadata.isSensitive) {
-        log.info("Deleting public record...")
-        em.remove(p)
-        // } else {
-        log.info("Updating public record...")
-        update(p)
-        // }
+        // delete the public record if it's become sensitive
+        if (o.getSensitiveRecord) {
+          log.info("Deleting public record...")
+          em.remove(p)
+        } else {
+          log.info("Updating public record...")
+          update(p)
+        }
       }
       case None => {
-        // create the public record if it's not sensitive
-        // if (!metadata.isSensitive) {
+        // don't create public record if it is sensitive
+        if (o.getSensitiveRecord) {
+          log.info("Not creating public record (sensitive=true).")
+        } else {
           log.info("Creating public record...")
           val p = new TaxonObservationPublic
           update(p)
           em.persist(p)
-        // } else {
-          log.info("Not creating public record.")
-        // }
+        }
       }
     }
   }
